Add files via upload
A Java program which merges two ordered singly linked lists of integers into a single ordered doubly linked list. The resulting list does not contain any duplicate elements. In addition to routine testing, it demonstrates special cases, for example, one or both lists are empty, there are elements which occur in both lists and traversal in both directions.
The complexity of the solution is also discussed.